# [PERF] mrp_workorder: batch write on button_finish