### Description:

Improves the performance of `button_finish` when confirming many work orders. Previously, the method would update each work order individually, leading to slow performance. This PR changes the way the write is executed to do it by batch, which is much faster, especially when work orders share the same manufacturing order.

### Benchmark (in 18.0):

| N° of work orders | Before | After |
|-------------------|--------|-------|
|              250  |    14s | 610ms |
|              500  |    53s |    1s |

### Reference:

opw-4957774