@robodoo 2025-08-20
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/223588.png)](https://mergebot.odoo.com/odoo/odoo/pull/223588)